refactor: introduce `precompile/interfaces` package
